rag-docs/backend/requirements.txt
flask
pydantic
python-dotenv
langchain
langchain-community
langchain-core
langchain-chroma
langchain-google-genai
chromadb
google-generativeai
gpt4all
sentence-transformers
faiss-cpu
tiktoken
pdfplumber
python-docx
chardet
PyMuPDF
tqdm
streamlit
requests
huggingface-hub
transformers
